refactor: Extract shared utilities and organize update_constraint_tests by constraint type
Eliminated ~70% code duplication by extracting shared constraint setup utilities
and organizing tests into focused modules by constraint category.

Changes:
- Created constraint_test_utils.rs with shared table setup helpers and assertion utilities
- Organized tests into 4 focused modules:
  * not_null_constraints.rs - NOT NULL constraint tests
  * primary_key_constraints.rs - PRIMARY KEY constraint tests
  * unique_constraints.rs - UNIQUE constraint tests (single and composite)
  * check_constraints.rs - CHECK constraint tests

Benefits:
- Reduced code duplication from ~400 lines to shared utilities
- Improved test organization by constraint type
- Easier to add new constraint test cases
- Better test isolation for debugging specific constraint violations
- Consistent helper patterns across all constraint tests

Closes #702
